+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Server
+#### Added
+- Token longevity can now be configured (in seconds). Default to 43200 (12 hours).
 
 ## [1.0.0] - 2021-07-22
 ### Server